+++ src/main/java/flab/project/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package flab.project.mapper;
 
 import flab.project.data.dto.model.Profile;
-import flab.project.data.dto.User;
+import flab.project.data.dto.model.User;
 import flab.project.data.enums.requestparam.GetFollowsType;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
