+++ src/main/resources/mapper/flab/project/mapper/user-mapper.xml
@@ -38,7 +38,7 @@
             <if test="updateProfileDto.userName != null">user_name=#{updateProfileDto.userName},</if>
             <if test="updateProfileDto.profileImgUrl!= null">profile_img_url=#{updateProfileDto.profileImgUrl},</if>
             <if test="updateProfileDto.selfIntroduction!= null">self_introduction=#{updateProfileDto.selfIntroduction},</if>
-            <if test="updateProfileDto.department!= null">department=#{updateProfileDto.department},</if>
+            <if test="updateProfileDto.departmentName!= null">department=#{updateProfileDto.departmentName},</if>
         </set>
         WHERE user_id=#{userId}
     </update>
